Fixes #7: Check how the "reissue-cookie" works if it actually triggers
Yes, my suspicion was correct: If we find a "looped" message, caught by
the "reissue-cookie", we'll stop - but that message will just be
"dropped on the floor", the session will be committed, and we will have
lost that message.

Fix: Reissue that last message before we stop. So, it will be attempted
one more time, but we'll stop the consume loop.

Note that this should really never happen, due to count limitMessages,
and FIFO-ness of queues: Say there was 300 messages on queue, and the
user hits "reissue all" - which then defaults to 300 as limit. Even if
these messages are insta-DLQed upon retry, they will come in behind
those 300. So when we're finished with the 300, we'd start consuming the
301st, which is the already reissued-but-insta-DLQ-again'ed message.
However, the limit would stop us before consuming that 301st message.
But, if the user chose 1000 as limit (even though there was just 300),
we could conceivably end up in such a looping situation. The reissue
cookie would then catch it - but that first message where we catch it
would be reissued again. Which is not a problem.

The history here is that I first conceived of the cookie, then it
occurred to me that the limit might fix it just as well, just simpler,
and more robustly. So then we ended up with both solutions; The cookie
has probably never hit.